more clenaups (mostly framework layer) of MSVC 2k17 W4 warnings - comment out unused params, a few extra casts on int type conversions etc

+++ Library/Sources/Stroika/Frameworks/NetworkMonitor/Ping.cpp
@@ -101,7 +101,7 @@ Pinger::Pinger (const InternetAddress& addr, const Options& options)
     // use random data as a payload
     for (Byte* p = (Byte*)fSendPacket_.begin () + sizeof (ICMP::V4::PacketHeader); p < fSendPacket_.end (); ++p) {
         static std::uniform_int_distribution<std::mt19937::result_type> sAnyByteDistribution_ (0, numeric_limits<Byte>::max ());
-        *p = sAnyByteDistribution_ (sRng_);
+        *p = static_cast<Byte> (sAnyByteDistribution_ (sRng_));
     }
 }
